Convert check email to ts (#2302)
* chore: let TS know that importing SVGs are fine

* refactor: check email page to TS

* refactor: check email page to TS
