+++ src/custom.d.ts
@@ -1,0 +1,4 @@
+declare module '*.svg' {
+    const content: any
+    export default content
+}
