+++ src/pages/check-email/index.tsx
